refactor(ml/): Rename to app.py and change running port number
