+++ ml/api/price_predicition.py
@@ -5,8 +5,8 @@ Routes
     - / : hompage
     - /api/price-prediction/{RARITY} : price prediction
 """
-from api import create_app
+from app import create_app
 
 if __name__ == '__main__':
     app = create_app()
-    app.run(port=8000)
+    app.run(port=4000)
